Optimising service level logic for some apis + little bit of refactoring.

+++ src/main/java/com/vistagram/app/api/response/PostDetailResponse.java
@@ -18,8 +18,8 @@ public class PostDetailResponse {
     private String poiName;
     private String poiLocation;
     private LocalDateTime createdAt;
-    private int likeCount;
-    private int shareCount;
+    private long likeCount;
+    private long shareCount;
     private boolean likedByCurrentUser;
 
     public static PostDetailResponse fromPostDto(PostDto postDto){
